requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cloudinary==1.41.0
cryptography==43.0.3
distlib==0.3.9
dj-database-url==2.3.0
Django==5.1.2
django-admin-honeypot==1.1.0
django-admin-honeypot-2024==1.0.0
django-admin-honeypot-updated-2021==1.2.0
django-allauth==65.0.2
django-cleanup==9.0.0
django-cloudinary-storage==0.3.0
django-environ==0.11.2
django-htmx==1.19.0
django-ipware==7.0.1
django-resized==1.0.3
filelock==3.16.1
gunicorn==23.0.0
idna==3.10
packaging==24.1
pillow==11.0.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
pycparser==2.22
python-ipware==3.0.0
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.26.6
whitenoise==6.7.0
